Refactor: split API per request type
