+++ pypuppetdb/api/__init__.py
@@ -1,0 +1,13 @@
+from pypuppetdb.api.query import QueryAPI
+from pypuppetdb.api.metrics import MetricsAPI
+from pypuppetdb.api.other import CmdAPI, MetaAPI, ServicesAPI
+
+
+class API(
+    QueryAPI,
+    MetricsAPI,
+    CmdAPI,
+    MetaAPI,
+    ServicesAPI,
+):
+    pass
