+++ tests/test_connect.py
@@ -3,7 +3,7 @@ import mock
 import pypuppetdb
 
 
-@mock.patch('pypuppetdb.api.requests.Session.close')
+@mock.patch('pypuppetdb.api.base.requests.Session.close')
 def test_connect_api(session_close):
     puppetdb = pypuppetdb.connect()
     assert puppetdb.version == 'v4'
@@ -12,7 +12,7 @@ def test_connect_api(session_close):
     assert session_close.called
 
 
-@mock.patch('pypuppetdb.api.requests.Session.close')
+@mock.patch('pypuppetdb.api.base.requests.Session.close')
 def test_connect_with_statement(session_close):
     with pypuppetdb.connect() as puppetdb:
         assert puppetdb.version == 'v4'
